Deploy artifacts to gitlab
We want later stages in the pipeline to be able to make use of our final
build artifact for deployment, additional testing, and so forth.

This patch also makes sure that we version our local artifacts in a way
that allows us to trace its origins easily (by way of git tag + hash and
a potential -dirty-marker).