ci: Add dockerized GitLab review apps
This infrastructure patch adds a new stage to our pipeline - namely,
a deployment to a review environment. The review environment consists of
dynamically created docker containers running the artifact produced in
the build stage. Each branch has its own environment, which is rebuit
and redeployed on every commit to that branch.

When merged into develop, this should give the nice side effect of
always hosting the latest version of develop at
rv-develop.gitlab.elvaco.local and master at
rv-master.gitlab.elvaco.local, though I haven't tried that out just yet
for obvious reasons.

Of course, this is not a replacement for a proper staging environment,
since the review apps don't have a "real" database with "real" data in
it.

Either way, this should allow us to more easily verify functional
changes (at least on an interface level), in addition to code changes
introduced by a given change as well as expose such changes to
non-technical stakeholders of the project.

The added CI.md file contains more documentation on the technical side
of this setup.