docs: Detailed architecture overview
The mvp_components diagram contains some of the components within Spring
that we use, and some of the most important front end components.

The diagram was produced in draw.io. draw.io supports modifying the XML
files, but I checked in the artifact as well. I don't want a build
procedure to involve visiting draw.io.

This fixes #4303

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>