ci: Turn down the review-app to use 64M+64M+.. memory
It seems hard to calculate the total amount of memory per Java
instance/thread. Decreasing XX:MaxMetaspaceSize and Xmx seems to impact
parts of the memory usage, but not make up for all of it.

When running multiple review apps in parallel, they need to be as cheap
as possible.

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>