Add PostgreSQL, liquibase, docker and docker-compose
In summary, this commit adds four new technologies to the project:
- docker,
- docker-compose,
- postgreSQL, and
- liquibase

Effectively, this patch adds a bunch of infrastructure changes to allow
a developer to run the application against a liquibase-created
PostgreSQL database in docker-compose.

Since our entities are all very placeholder'y and work-in-progress'y,
I've settled for only making one such entity (namely, dashboards)
available as a PostgreSQL table when the application is run with the
"compose" profile. The other entity componenents are limited to the
default profile.