Add spring-boot-devtools
spring-boot-devtools enables us to speed up our development cycles
significantly, without having to rely on a full-blown class-rewriting
hot reloading technology such as JRebel.

It works by running two separate class loaders, one for classes that
don't change, and one for classes that do. When a change is detected
(typically as a result of a build), only the latter classloader needs to
be restarted, which improves restart times a great deal.

Developer tools are automatically disabled in "production" mode; i.e
when the application is launched using `java -jar`.

More docs at:
https://docs.spring.io/spring-boot/docs/current/reference/html/using-boot-devtools.html