Introduce integrationTest gradle module
The tests are separated with unit tests and integrations tests. Unit
tests are located in 'test' directory and integration tests are in
'integrationTest' directory.

If you want your test to run against spring boot application in your
integration test, please extend the abstract 'IntegrationTest', e.g.
this is the class that starts up the spring boot test runner.