poc: Properties saveable as jsonb
This patch adds a PoC-implementation of properties on JPA entities that
are stored to as jsonb (in PostgreSQL) or regular json (in H2),
depending on the database backend used.

Right now, the API is pretty anemic - we only support storing either
lists-of-int or strings, and only on the top level of our root object.

Again, only proof of concept.

This embodies the outcome of the investigation carried out as part of
issue #4344.

As a bonus we now also have .editorconfig support for *.gradle files!
(read: a merge conflict messed up indentation, and I lazied out and
fixed it in this commit instead of splitting it up and fixing it
properly.)