Tidy up gradle tasks
As the comment says:

"
The bootRepackage task creates a fat jar of the project and, by default,
excludes devtools if it exists. However, it depends on the distTar & distZip
tasks which are provided by the 'distribution' plugin (which is, in turn,
applied implicitly by the 'application' plugin). This means that we can't
make use of the artifact produced by distTar directly and have that include
the fatjar. This is supposedly fixed in Spring Boot 2.0.0M1 Instead, we need
the fatjar (and startscripts), or disable bootRepackage altogether and
manually exclude devtools.

The latter feels less intrusive and less involved, so let's go with that.
"