idea: Share _some_ IDEA settings
It feels a bit draconic to not allow sharing settings at all. For
example, run configurations are an excellent candidate for sharing
between setups.

We only need to be careful to not share user-specific settings, and to
always make sure that we do not rely on the IDE to perform some task.

This patch incorporates parts of
https://github.com/github/gitignore/blob/master/Global/JetBrains.gitignore

into our .gitignore, and adds a run configuration "MvpApplication:
Postgres in Docker", that I believe should benefit most of us once
http://gitlab.elvaco.local/elvaco/mvp/merge_requests/26 is merged.