ci: Deploy all commits for develop onto staging server
staging.md contains a detailed list of manual steps to take, when
preparing a staging server.

This is what happens:
- Gitlab runs our pipeline target called "deploy_staging", in the
  "deploy" phase
  - This is done within a docker container, so take all necessary
    variables such as ssh keys that allows you to push to the server,
    from the Gitlab's "secret variables"
- Some verification of the secrets stored in Gitlab is done
- Gitlab deploys the .tar artifact and a corresponding systemd service
  file, and restarts the service on the "other side", i.e. the staging
  server

These are the design choices made on the server:
- Run the service as the unprivileged user (UP) "elvaco_mvp"
- Use authbind to allow the UP to listen on port 443
- Use sudoers rules to allow us using the UP when deploying artifacts

This should minimize the attack surface a bit, since we never use root
directly.

The systemd file is very small so far, I expect it to change quite a
bit. For example, Restart=Always may be a thing we want.

Hopefully, we can reuse all of this for deploying to the production
server, in the future.

This fixes #4302

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>