Refactor to handle login when refreshing the page
It the user is logged in and the page is refreshed, the the user should
not be logged out. The auth state is persisted in the local storage
until logout. This might be changed to time based login in the future.
Now the user will be logged in until logout!