Add i18next-react translation functionality
First the POT-file is created by extracting all translation keys from the code,
identified by <div>{translate('hello')}</div>.  Here is the key 'hello'.

The POT-file is used to create different translations and saved in PO-files in
src/i18n/locales.

The PO-files are then converted to JSON-files and read in by the
i18next-framework.

This fixes ticket #4276