ci: Make version file an artifact for review-deploy
We need to be able to figure out which artifact to deploy for review.
Previously, we used `git describe ...` to figure out what version we
were on, but this (obviously, duh) doesn't work with GIT_STRATEGY=none.

This patch simply adds a task to write down the version in a file
(./.version), records that file as an artifact in the build step, and
uses its contents to determine which distribution archive to deploy for
review.