frontend/config: Evaluate configuration lazily, add types
When the application had been deployed to mvpstaging.elvaco.se, all
requests went to /dashboard/current instead of /api/dashboard/current.
The reason for that was that using a hard coded `location.host` in the
configuration, made no difference, since the "dynamic" configuration was
evaluated when building instead of in run-time.

Also.. damn, typing is nice.

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>