frontend/config: Add tests for detecting env
There were a couple of cases that the type checking of TS did not help
us enough. Adding end-to-end tests makes us comfortable with determining
that "a config" (a set of configuration options) is correctly obtained
based on the global node variable `process.env.NODE_ENV`.

The configuration files are moved from src/ to a use case as well, for
namespacing purposes.

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>