Add initial version of rabbit-consumer
This is a service which consumes meter messages conforming to the Elvaco
MVP MQ message schema from a RabbitMQ queue and publishes them to a
PostgreSQL database which exposes an add_measurement(...) procedure.

In this version, a lot of settings (such as queue and database
settings) are hardcoded that shouldn't be, and what logging there is is
done directly to stdout instead of using a proper logging framework. In
other words, this is far from being production ready. However, I think
it is a good starting point.

This patch also includes two kinds of test data producers - one which
randomly generates messages, and one which simply publishes the included
message-samples.