table: Feed data from <Table> to its children
The pattern `<Table><Column/></Table>` was inspired by
[Facebook's Fixed data table](https://github.com/facebook/fixed-data-table)
(which was deemed a bit big to pull in completely). Dealing with
children of React components is explained well in
[A deep dive into children in
React](http://mxstbr.blog/2017/02/react-children-deepdive/). This
pattern also requires turning off the tslint option "jsx-no-lambda"
which is fine in this particular case because we're not looping over
the callback (don't do that!)

Since the previous try, the data has been made too look like the output
from [Normalizr](https://github.com/paularmstrong/normalizr), based on
[Normalizing State Shape](http://redux.js.org/docs/recipes/reducers/NormalizingStateShape.html)
from the Redux guide.

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>