DashboardContainer: Move callbacks out of JSX element's attribute
jsx-no-lambda is a good thing to keep in the jslint configuration.

The callback does not depend on the current iteration of the parent
element, i.e. it is pure (fed all arguments that makes up for its
state), so we can extract it to avoid recreating the same thing over and
over.

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>