db/liquibase: Create users table
One couldn't login when starting the "MVPApplication: Postgres in
Docker" Spring Boot configuration. The exception said that the users
table didn't exist, so I created it. In the future, it will hold
password hashes and salts, but that's for another iteration.

"An example online" used the id field of the db changelog YAML as an
incremented value. I think that looked like a good thing, so I changed
the previous DB changes to that scheme.

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>