frontend: Deploy frontend only at mvpstaging.elvaco.se:3000
In preparation for _the November demo_, we decided to skip the Java
bits, for faster iterations GUI-wise. Instead of messing with the
regular installation, we will deploy another artifact and host it with
JSON server's built-in web server.

The port is not open quite yet, meaning that it doesn't work end to end
yet, but I walked through the deployment process manually and corrected
all the kinks I found.

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>