docs: Move staging.md to deploy/, extract sudoers rule to own file
It is pretty pleasant when you browse the repository through Gitlab and
you get a short blob of text related to the directory you just navigated
to. Just as you did for frontend/ forever, you now get an overview of
the things in deploy/ as well.

By creating a separate file for the sudoers rules, we can deploy it as
is.

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>