frontend: Use dynamic mock data
It's really annoying to maintain a JSON file with thousands of entries
if you want to, say, add an extra column to all of them.

Instead, we let JSON server host the API through its built-in Express
server, so that we can randomize data. The randomized data is seeded
(with the help of "seedrandom"), which means that as long as we don't
change the generator, all of us will see the same data. Everybody wins.

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>