Add geocoding capabilities to mockdata.js
This patch makes it possible for mockdata.js to fetch coordinates from
geocoding.xyz for meters and gateways read from the seed data files. The
results are cached in data/geocoding.json to avoid repeated work.

The geocoding.xyz service is a rate-limited API, so we limit our
requests to one per second to avoid being throttled.

Currently, the feature is disabled by default, and the inital geocoding
dataset only covers a miniscule subset of the actual addresses. We
should do a proper run of all the addresses in our seed data and commit the
result, shortly.