Add initial JPA wiring for measurements and meters
This patch adds repositories, entities and controllers for
measurements and physical meters.

It also includes integration-test support for fetching paged generics
through REST, which wasn't previously possible.

A couple of our endpoints will be paged - no getting around that. We'll
want to be able to reflect this in our integration tests.

Allowing us to write pretty tests requires a bunch of type hacking to
work around Java's type erasure and still get the results we desire, so
this part might look a bit more involved than you might expect.