Add measurement data to mockdata.js
This code is pretty awful to look at, but as the inline comment says,
it's necessary.

The libraries that I found for parsing CSV to JSON were way, way too
slow to provide an acceptable startup time with the amount of data that
we're working with (we're talking days ...).

So this'll have to do - ugly but functional. Story of my life.

Oh! And you _don't_ want to open this endpoint in your browser without
pagination. Trust me. Use /measurements?_page=<N> instead. You've been
warned.