Remove LogicalMeter
This should probably not be a directly accessible object - the metering
point itself should be represented by a MeteringPoint (to be added), and
the meter definition (i.e, what quantities are present for a given type
of meter, etc,) should be represented by a MeterDefinition object.

At least that's my thinking right now, so I'm cleaning this cruft up for
now and we can readd it if it turns out that I'm crazy.