Make paging work for both H2 and PostgreSQL
Paging, that is. Sorting is still very questionable if it works, at
all. However, I do believe that in all reasonable cases, measurement
data should always be retrieved as a series, so it might not be too
bad.

Spring 2.0.0 documentation hints that paging is possible with native
queries by simply providing the countQuery to the @Query annotation:

"Note, that we currently don’t support execution of dynamic sorting for
native queries as we’d have to manipulate the actual query declared and
we cannot do this reliably for native SQL. You can however use native
queries for pagination by specifying the count query yourself"
--
https://docs.spring.io/spring-data/jpa/docs/current/reference/html/#_native_queries

That doesn't work in any of the 1.5.x versions, though, afaict.

Might be worth looking into further if this approach turns out to be
unsuitable at some point.