Add querydsl model generation and unify output dirs
Mostly prep work for the following commit, just to keep things tidy.

This commit also consolidates the output directories of gradle and
Intellij IDEA instead of having the latter put its build artifacts in
out/ whereas gradle would put them in build/.

This used to be the default behaviour in IDEA, but changed a while back
in response to an issue report regarding multilingual projects -
something that does not affect us. I find the behaviour implemented by
this change much more intuitive.