tsx: Always use "import './SomeFile.scss'"
The ./ in ./SomeFiles.scss causes the file to be properly referred to if
the file exists in the same folder. If we leave it out, and try to

	import 'SomeFile.scss'

then tslint will warn us about "SomeFile.scss" not being a listed
amongst the other dependencies in package.json.

This may be new behavior since the recent upgrade of tslint.

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>