Enable checkstyle for backend
This commit adds the checkstyle gradle plugin together with a
configuration that makes sure that the code is formatted according to
Googles Java Style Guide
(https://google.github.io/styleguide/javaguide.html), with one
modification.

We diverge from the above guide in that we do not require javadoc on all
public methods, since we believe that descriptive, easily-read method
names are more important, and that we will be able to catch methods
where javadoc is necessary in review instead.