Remove obsolete loaders
These loaders were nice to have once, now they're mostly in the way in
that they load data that becomes accessible for integration tests and
such.

We should decide how we want demo/seed/test data to be loaded, what data
to load from where, and under what conditions eventually, and implement
that properly instead of this interim hack.