css: Use .first-uppercase instead of capitalize
Pre this patch, a tab would be called "Hello There" instead of "Hello
there".

We cannot reuse the cute .first-uppercase class in the
SideMenuContainer's AppTitle, because it creates its <h1> itself. It
takes a title property, instead of letting us pass our own <h1> as the
elements children.. so it goes.

We can also remove the .capitalize class, since it is unused.

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>