period: Remove "custom"
Our mapping between Period (fixed values) and custom is not complete as
of now. Instead of making it more competent, I bail out. At least some
of us have wished for a single calendar control, instead of two
combined, so the previous control that used two calendars was not good
enough.

The next "custom date picker" should have the ability to select time as
well as date.

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>