period: Test startAndEndFrom by extracting "now" and using curry
This means that we want to either wrap a function; use bind() and rely
on this; or import a curry implementation like the one found in Ramda. I
opted for the latter, because I think JS's bind() makes it too hard to
keep track of the current scope, and currying keeps the code really
short and simple.

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>