Add zuul, netflix's jvm based router
Also use *.yml instead of *.properties files. Intellij is providing you
with much better autocompletion for spring profile, when using yaml
instead of properties.

All the user and authentication endpoints are now requesting java rest
api, and the other endpoints that are not fully implemented in the backend
are just routed to json-server with zuul.