Move UserMapperTest.java to base module
This test class should not reside here in this module in the future.
However, I could not solve the gradle build dependencies error. When the
whole project is build the dependencies to core module from database
test classes is not working - the classes in core module cannot be found.

We must solve this later!