Drop support for runFinalJar
I don't think anyone ever used it, since it's easier to just run
everything through the IDE. Besides, it relied on the spring-boot
Gradle plugin to assemble a fatjar - something that's not done anymore.

Additionally, the artifact to use to run check-off tests is the
distribution archive, and that's used in the compose case, already.