Don't run frontend tests twice
The check task in build.gradle already has a dependency on 'yarn_test'.
We run check when building, so the tests are already run. The removed
'tests' target is defined in fuse.js and does pretty much the same thing
as 'yarn test', but not really.

Anyhow, it should be enough to run the tests once.