report: Plot graph with real values from /api/measurements
We collect values from the paginated endpoint of /api/measurements but
we must add support for querying for a specific meter, and for
subsequent pages of measurements.

Calling the REST API is quite a mess at the moment. We should not couple
the input/output format of the data sent to/received from the REST API,
with the function that performs the HTTP request (which responsibility
is to send data and know what an HTTP error looks like, etc).

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>