Auth /api/user based on principal organisation
This patch adds an 'OrganisationPermissionEvaluator' that checks the
current user's roles and organisation to verify that they have access to
the kind of operation that they are trying to perform.

For now, we only validate UserDto access on REST endpoints, but we'll
probably want to add handling for other types as well, shortly.

The code in the OrganisationPermissionEvaluator is a bit messy, because
I didn't want to spend too much time over-architecting it before we're
sure we want to go down this route.

This fixes MVP-138, MVP-137, MVP-135, MVP-166 and MVP-134.