Add Settings and make demo data loading "idempotent"
This patch adds a new domain model (and scaffolding) for storing
settings, which is used to check whether demo data and demo users are
already loaded and only loading those otherwise. This allows us to
restart the application "idempotently", without tearing down the
database between runs (as long as you don't change any
data between runs).

The Setting model is fairly trivial, basically just a string->string
mapping, but could be extended in the future to contain descriptions and
other value types as well, as we see fit.