Basic authorisation checks for measurements
This patch makes sure that only users that should be able to see a
certain measurement are able to do so.

Since this endpoint is paginated, we can't use the PostFilter
annotations, because that would filter out the already-paginated
results, leaving us with confusing, inconsistent results.

Instead, we access the security context directly and modify the
query-to-be-executed according to the current principals authorisation.

This should probably be extracted into a separate module/layer/package
and consolidated with the existing, annotation-based, authorisation
strategy in the near future to keep things neat and tidy.

I also did some refactoring in the integration test classes to extract
the convenient 'asElvacoUser(), asSuperAdmin(), ...' methods to the
super class.

This fixes MVP-168.