Authorise user access in core
As a side note, since we make use of UserUseCases in our
UserDatabaseLoader, we also need to manually set an Authentication the
current context, there. Another consequence is that we cannot rely on
Session scoping of the currentPrincipal bean, since we're using it
outside of a HTTP context - instead we use the standard Prototype scope
for said bean.

I also introduced a separate OrganisationMapper, since we had some
duplicate code that fit well in there, and I didn't want to add even
more of that.

Mostly, though, just moving stuff around and relying less on Spring
Security for our authorisation needs!