login: Window (about 500px height) moves login out of view
When the login screen is really small (often seen if you have the
developer tools docked to the bottom), blindly adding 15% bottom margin
pushes the login form's container out of view. The golden ratio may look
better, but we need to solve it in another manner.

We're not trying to conform to responsiveness, but this jumps in your
face and can be easily avoided.

Signed-off-by: Carl Helmertz <carl.helmertz@elvaco.se>