gradle: Set up and tear down DB for 'systemTest`
This patch adds rudimentary setup and teardown of the postgresql
database when the `systemTest` task is run. It could be vastly improved,
for example by picking an available port at random, to avoid thrashing
an already running instance, but at least it's a step in the direction
of making the systemTests more convenient to run.