Rename MeteringPoint to LogicalMeter
This is a huge refactoring commit that should contain no functional
changes - just renaming.

The rationale for this rename is that we seldom talk about "metering
point", instead preferring either "meter" or "logical meter". The meter
model also contains information that is not strictly related to its
physical whereabouts, such as medium, making the metering point name
feel a bit awkward.