ITs: Remove dependency on UserDatabaseLoader
Many of our integration tests depend on UserDatabaseLoader having run
prior to execution. This is, of course, slightly insane.

For convenience, I've consolidated the creation of the most commonly
used users into the base IntegrationTest class. Creation (and deletion)
of entities/objects that are only relevant to a specific test has been
moved to that test's fixtures.