Add externalId to physical meter, rename identity -> address
This change adds an "externalId" field to the physical meter model. This
is an externally allocated identity of the metering point that this
meter is situated on - likely some sort of a facility identifier. In
other words, this identifier serves as the link between a physical meter
and its logical counterpart in the system.

To avoid having too many "identifiers", this change also renames the
physical meter's "identity" to "address". This field represents
the meter's identity from its associated gateway's point of view -
usually, the identity will be some sort of address (such as an M-Bus
secondary address).

In addition, a new method is introduced in the PhysicalMeters interface
to find a meter based on its owning organisation, its external ID and
its address. These are the three parts we need in order to uniquely
identity a physical meter in the system. I've removed the old
findByIdentity, since it was insufficient for its intended purpose.