Require meter definition for logical meter
This change makes the meter definition on a logical meter non-nullable.
Of course, it's not necessarily the case that we know which meter
definition we should apply for a logical meter at all times. This
condition is represented by a new sentinel "null object", called
UNKNOWN_METER.

This lets us get rid of a bunch of messy null checks, and delegate the
handling of the "unknown meter type" case to the meter definition
itself, as necessary.