Avoid leaking data through unmapped exceptions
Error messages are good, but only when they're actually intended for
human consumption. This change causes exceptions that are thrown during
a request to be returned as internal server errors (HTTP 500) with a
hard coded message if no explicit mapping exists for them, either as a
separate handler implementation, or as a @ResponseStatus mapping on the
exception type.

We do this to

a) Avoid confusing the user with technical details that they should not
have to see, and
b) To avoid leaking potentially sensitive data that might be present in
the exception message.

In summary; be less promiscuous in what we expose to the end user.