Fix basic-auth internal server error bug
This patch fixes an issue where accessing an endpoint other then
/authenticate using basic auth would trigger an internal server error.
This was caused by the MethodSecurityConfig not taking into account any
other form of authentication than the token bearer type.

This, in turn, triggered a BeanCreationException for the currentUser
bean, which was then (correctly) propagated to the user in the form of
an internal server error.