test: Add factory class for easy context management
This patch adds a factory which is able to construct an
IntegrationTestFixtureContext. The context is intended to isolate a
single test case, by generating a new organisation. This is an approach
recommended by the "Continuous Delivery:  ..." book, since it reduces
the risk of inter-test-case pollution, and also trivializes
parallelization of test cases.

For now, the context only stores an organisation and an associated
regular user, but we might want to flesh this out eventually with more
sophisticated context construction methods to reduce the amount of
boiler plate in our tests.

As an added bonus, move "isPostgresDialect()" to base IntegrationTest
class, since I'll need that again in a little while.