Use separate database user for application
This is a security & deployment fix. We don't want our application to
accidentally go all Bobby Tables on us, and we don't need it to be able
to, since all schema migrations should be performed by liquibase
anyways.

This way, we're at least safe against that particular type of data loss.

Needless to say, the hard coded password should not be used in
production, but is only for demonstrative purposes.