docs: Delete out of date MQ protocols
The protocols live in the code, within the rabbitmq-consumer module.
It's better to not have any formalized documentation, than documentation
that is wrong. Another option would be to start using the message
protocol defined here, but we have already started to parse the JSON
ourselves.

Feel free to read the test code in order to understand the messages.