RabbitMQConsumerTest: Assert meter is created
We have a race condition in this test: A foreign key constraint from
logical meters to organisations causes the test to fetch when we try to
delete the organisation, despite deleting all logical meters before.

If this doesn't work, there's something wonky with the way Hibernate
decides to flush deletes.