Measurements endpoint takes logical meter IDs
This patch changes the /api/measurements endpoint to work on logical
meter IDs, instead of physical IDs. From an end-user viewpoint, this
should be a more common use case, since the logical meter is what is
most commonly exposed to the user.

This patch also makes it possible to specify multiple quantities, and
scale them individually, which is useful in the case where we want to
plot/graph multiple quantities for multiple meters.

I've preserved the behaviour that all parameters are optional, even
though they are now listed explicitly, and added a few missing tests.

The endpoint currently performs one query to the use case (and thereby
to the database) for each quantity, so that could be optimised in the
future.