Rename organisation code to slug and add external ID
The 'code' is really a slug. This patch renames the field to reflect
this, and adds a convenience class for generating slugs from
organisation names (or other strings).

We also add an externalId field, which serves the same purpose as the
respective field for meters, namely as a reference for external systems
which add data to the system.