Don't try to create same organisation twice
This change fixes a bug where the ProductionDataLoader would try to
re-create the initial organisation, even if it existed (this was
probably why we used to have a "is production data loaded" check here
previously).

The actual fix is just two lines of code, the rest of the changes are
unit tests and refactoring to make it possible to unit test the data
loader.