sql: Key tenant-owned entities by organisation
This patch changes the primary keys for entities that are owned by an
organisation, to include the organisation ID.

I was going to reflect this mapping in our JPA entities as well, but
that proved *much* harder than I had hoped, and after a day's worth of
blood, sweat and tears trying to get it to work I had gotten nowhere.

Anyways, the original reason we wanted to do this *now* was to avoid
having to migrate primary keys in the future, and this patch achieves
that, at least.