rabbit: Handle measurement updates
When a measurement exists for a meter, at a time which we already have,
we now update the existing measurement with the unit, quantity and value
of the new measurement. This is in contrast to the old, less helpful
behaviour, where we would throw an exception due to database constraint
violations as a result of trying to insert a new measurement for the
same meter at the same timestamp.