rabbit: Build responses for mesasurement messages
When we receive a measurement for a structural entity
(gateway/meter/...) that we currently don't know anything about, we want
to publish a "get reference info" message for Metering to consume. If
Metering has more information about the object, it will eventually
publish this information in the regular way for us to complement the
data on our end.