measurementActions: Trim average if before first measurement
The API is friendly enough to give us average values for the hour 9 if
we ask for data from 9:15 and later on. The rendered graph looks really
bad with only average values and not individual values, though, so we
remove the early birds. IMO the API is good enough, this is a rendering
issue.