metersWithinThreshold: Handle undefined
We could argue that metersWithinThreshold() should only accept sanitized
values but since the type system couldn't help us here, I don't think we
should feed bad about being defensive in this case.

Without this patch (or something else that checks .entities before
passing it to metersWithinThreshold()), the collection page errors out
if the user has selected parameters enough to not get any hits, i.e. has
an empty selection.

This bug was introduced by me, when I translated the method to return a
list of items instead of a boolean, and I didn't manage to trigger this
particular state.