Don't create multiple entities for same gateway
When we've received a gateway reference, and subsequently receive
updated information about that gateway, we would previously create a new
entity for it. This is not the desired behaviour, we should instead
update the existing gateway with whatever new information is now
available to us.

This patch makes it so.