Don't assume uniqueness for measurement timestamp
A measurement is unique in (meter-id, quantity, timestamp). The quantity
part is important, since we normally get multiple quantity values for
the same timestamp.

Of course, this makes it impossible to update an existing quantity the
way we've done it, but that's a much rarer event (if it ever
happens?), so let's support the normal case instead.