rabbit: Map meter statuses
When a message is received, with a new status for a meter, we want the
status to be updated on the meter. If the same status as the already
active one is received, we ignore it.

We also make sure to stop other active statuses on the meter when a new
status is received.