rabbit: Fix rejection requeuing
Apparently, Spring thinks it's a good idea to always requeue a message
if the receive method was invoked with an external transaction manager.
This means we have to set _another_ property in order to really reject
(and not requeue) invalid messages.

I also took the time to move the configuration for rejection requeuing
to the consumer properties block, instead of using springs rabbit
listener namespace, since this makes things a bit more approachable.