Define a default presentation mode for quantities
This is a rather large refactoring, which adds (but does not make
effective use of) a definition of a "default presentation mode" for
quantities, to be used when retrieving a series for a meter.

The presentation mode can either be 'readout' - in which case each value
is presented "as-is" (i.e, the raw measurement value ) - or
'consumption' - in which case we present the value as the difference
between this value and the last value received for this meter & series.