CI: Use default salt branch when on develop branch
There seems to be some kind of protected keyword 'develop' thus we don't
find any files to use in our deployment.

By not setting saltenv when we are on develop, we use the default
salt has, which is develop for our configuration.

This will make it possible to release both automatic from develop
and via a git tag.

Signed-off-by: Mattias Ryrlén <mattias.ryrlen@elvaco.se>