Update redux-persist, to support migrations
The bad commit is 8004d71c, so 8004d71c~1 is a good place to get the
current localstorage from.

We cannot reuse the already broken state, because redux-persist used to
prefix all localStorage keys with "reduxPersist:" and now the keys are
prefixed with "persist:". There is a configuration option for overriding
the prefix, but the redux-persist/src/types.js says this:

    ...
    export type PersistConfig = {
      ...
      keyPrefix?: string, // @TODO remove in v6
    ...

On the upside, I could not break the state like it used to.

When I manually:
- modified a whitelisted part of the state (I added a new property to
  "auth")
- persisted it (by logging out and closing/reopened the window)
- modified the new property's type
- reloaded the application

then I found that the property I had in my localstorage, was replaced
with the initialstate, which of course contained a well-typed value.