api: Return 400 Bad Request for invalid quantities
This patch fixes a bug that would cause an Internal Server Error when
the /measurements/average endpoint was called with a mix of valid and
invalid quantities for a given meter.

I've discussed the desired behaviour with Erik W. and we arrived at the
conclusion that asking for invalid quantities is a user error at the API
level. In the UI, it should not be possible to trigger this error (even
though, at the time of this writing, it currently is).