Introduce toggle(item, list) which adds to, or deletes from, list
Note that toggle() does remove non-unique items, due to the
implementation. Since it's only used on "lists with IDs" from normalizr,
this should have no effect.

Frankly, I guess that normalizr use lists and not sets because of sets
not being mainstream enough. There is no need for the IDs to be in
order, or for there to be duplicate IDs, which are the selling points of
using arrays instead of sets.