ci: Pipeline tweaks
* Run fast and often-failing tasks earlier
The pipeline's job is to reject a build as soon as possible, since this
reduces the feedback cycle time and ties up fewer resources. It's a
waste of time to start up upgrade-, integration-, and system-test for
a branch which has trivial style issues, for example.

* Split up build of frontend and backend
These are not interdependent. The 'assemble' stage takes care of
packaging stuff up into the final dist-tar.

* Cache more things to avoid unnecessary rebuilding
Gradle keeps its task history in the .gradle directory, which wasn't
previously being cached.

* Define inputs/outputs for tasks to allow incremental builds
This avoids unnecessarily rerunning yarn_install over and over again.
Not very costly, but pointless.

* Only run findbugs on main source sets
We don't really care about bugs in our test code for the sake of the
pipeline, and findbugs takes a long time to run for our test source
sets.